% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVboostQscore.R
\name{DVboostQscore}
\alias{DVboostQscore}
\title{Generate Q scores for all SVs}
\usage{
DVboostQscore(DVb.res, tmp.mtx)
}
\arguments{
\item{DVb.res}{output generated by \code{\link{runDVboostwrapper}}}

\item{tmp.mtx}{output generated by \code{\link{loadVariants}}}
}
\value{
A data.frame with same format as \code{tmp.mtx} with one additional field \code{Qscore}
}
\description{
After training procedure, DVboost assigns Q scores to each SV based on training on a subset of SVs (only DEL). The higher the Q score, the more likely it is a true SV.
}
\examples{
data(ExampleData, package='DVboost')
sample <- 'NA12878'
outdir <- getwd()
tmp.mtx.DEL <- ExampleData[ExampleData$SVType == 'DEL',]
truth.vec <- tmp.mtx.DEL$CNVMAP == 1 | tmp.mtx.DEL$CNVR ==1
is.semi.truth.vec <- as.numeric(truth.vec)
DVb.res <- runDVboostwrapper( var.atr.mtx = tmp.mtx.DEL, var.ID.vec = rownames(tmp.mtx.DEL),
                             is.known.var.vec = is.semi.truth.vec,
                             output.DIR.name = outdir, input.sample.ID=sample, bySVlength=FALSE)
outmat <- DVboostQscore(DVb.res, ExampleData)

}
\seealso{
\code{\link{runDVboostwrapper}}
}
