% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDVboostmodel.R
\name{fitDVboostmodel}
\alias{fitDVboostmodel}
\title{Fitting a DVboost model based on annotated SVs}
\usage{
fitDVboostmodel(input.mtx, is.known.variant, fitting.verbose = FALSE,
  min.N.known.var = 50, sampleid)
}
\arguments{
\item{input.mtx}{annotated SVs loaded and formatted via \code{\link{loadVariants}}}

\item{is.known.variant}{numeric vector of 0/1 in the same order as \code{input.mtx} indicating whether a SV is known(1) or novel(0)}

\item{fitting.verbose}{logical value, if TRUE, it will print out progress and performance indicators. Default to FALSE}

\item{min.N.known.var}{minimum number of known SVs needed to build training model. Default to 50}

\item{sampleid}{same sample name used for annotation procedure}
}
\value{
DV.res which is essentially a \code{\link{gbm.object}} with several additional fields:
\itemize{
\item fitted.values: converted to probability based on \code{fit} field
\item ID: supplied IDs for SVs
\item is.known.variant: 0/1 indicating whether the SV is known(1) or novel(0)
\item DVboost.Q.score: Q scores for SVs
}
}
\description{
Internal function called by \code{\link{runDVboostwrapper}}
}
\examples{
data(ExampleData, package='DVboost')
sample <- gsub('_PE', '', grep('_PE$', colnames(ExampleData), value=TRUE))
outdir <- getwd()
tmp.mtx.DEL <- ExampleData[ExampleData$SVType == 'DEL',]
truth.vec <- tmp.mtx.DEL$CNVMAP == 1 | tmp.mtx.DEL$CNVR ==1
is.semi.truth.vec <- as.numeric(truth.vec)
DV.fit.res1 <- fitDVboostmodel(input.mtx=tmp.mtx.DEL, is.known.variant = is.semi.truth.vec,
sampleid=sample)

}
\seealso{
\code{\link{runDVboostwrapper}}
}
