% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVariants.R
\name{loadVariants}
\alias{loadVariants}
\alias{ExampleData}
\title{Load annotated SVs file}
\usage{
loadVariants(input)
}
\arguments{
\item{input}{input file after annotation procedure}
}
\value{
formatted data.frame to be used for model training including following fields:
\itemize{\item ChrA: first chromsome number;
\item Start: start position of first chromsome;
\item ChrB: second chromsome number;
\item End: end position of second chromsome;
\item SVLen: length of SV;
\item Germline: integer indicating germline artifact;
\item SVType: character value indicating SV type, could be one of the following 'DEL', 'DUP', 'INV', 'BND';
\item PE: number of paired end reads support;
\item SR: number of split reads support;
\item CNVMAP: 0/1 indicating whether the same SV is a known SV based on polymophic CNV database (https://www.nature.com/articles/nrg3871);
\item CNVR: 0/1 indicating whether the same SV is a known SV based on 1k genome database (https://www.nature.com/articles/nature15393);
\item avgL: average of strict mask L (depth of coverage is much lower than average). All strict mask averages are calcualted using start (100bp) and end (100bp) regions of the SV
\item avgH: average of strict mask H (depth of coverage is much higher than average).
\item avgZ: average of strict mask Z (too many reads with zero mapping quality overlap this position).
\item: avgQ: average of strict mask Q (average mapping quality is too low).
}
}
\description{
After annotation procedure, SVs are then loaded and formatted properly for model training
}
\examples{
data(ExampleData, package='DVboost')
### example data loaded and formatted with loadVariants
str(ExampleData)
}
