% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricView.R
\name{metricView}
\alias{metricView}
\title{Internal function called by \code{\link{runDVboostwrapper}}}
\usage{
metricView(DV.fit.res, input.sample.ID, plot.filename,
  DV.sampleQC.filename, subset)
}
\arguments{
\item{DV.fit.res}{output generated by \code{\link{fitDVboostmodel}}}

\item{input.sample.ID}{same sample ID used for annotation procedure}

\item{plot.filename}{full path to the figure file regarding training performance}

\item{DV.sampleQC.filename}{full path to the text file that contains metrics for model training performance}

\item{subset}{a vector of integers indicating a subset of the SVs that all metrics will be evaluated on}
}
\value{
\describe{
\item{A figure file}{The figure is saved at \code{plot.filename} and contains 6 subpanels, including \enumerate{\item distribution of known(1) versus novel(0) variants; \item feature importance for model training; \item distribution of Q scores for known and novel variants; \item precision, recell characteristics and best F score; \item ROC curve; \item FDR versus Q score}}
\item{QC metrics file}{The .txt file is saved at \code{DV.sampleQC.filename} and contains 6 fields, including \itemize{\item sample.ID: input sample ID \item Fscore_PR: F score for precision&recall \item AUC.PR.known.var: Area under the curve for precision and recall \item AUC.FP.TP.known.var:AUROC \item npos:number of known variants used for training \item nneg:number of novel variants used for training }}
}
}
\description{
After model training and testing, \code{metricView} generates summary statistics and figures for users to examine the training performance
}
