% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverallStats.R
\name{plotOverallStats}
\alias{plotOverallStats}
\title{Generate overall performance metrics for multiple samples}
\usage{
plotOverallStats(outputdir, type = c("Both", "Fscore", "AUROC"),
  bylen = FALSE)
}
\arguments{
\item{outputdir}{full path to output directory as specified in \code{runDVboostwrapper}}

\item{type}{specify which type of metric is preferred, could be one of 'Fscore', 'AUROC' or 'Both'}

\item{bylen}{logical value indicating whether to examine the performance by SV length instead. Default to FALSE}
}
\value{
a ggplot object
}
\description{
If you are running DVboost on a set of samples and output to the same directory, the you can use \code{plotOverallStats} to examine the overall
performance for all samples
}
