% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDVboostwrapper.R
\name{runDVboostwrapper}
\alias{runDVboostwrapper}
\title{Core function for DVboost model training}
\usage{
runDVboostwrapper(var.atr.mtx, var.ID.vec, is.known.var.vec,
  output.DIR.name, input.sample.ID, fitting.verbose = FALSE,
  min.N.known.var = 50, bySVlength = FALSE)
}
\arguments{
\item{var.atr.mtx}{a data.frame containing subset of total SVs (Only 'DEL' type)}

\item{var.ID.vec}{a vector of characters indicating IDs for each SV in \code{var.atr.mtx}}

\item{is.known.var.vec}{a vector contains 0 and 1s in the same order as \code{var.atr.mtx} indicating whether a SV is known(1) or novel(0)}

\item{output.DIR.name}{path to output folder where all figure files and data files will be saved}

\item{input.sample.ID}{character value used as prefix for output files}

\item{fitting.verbose}{logical value, if TRUE, it will print out progress and performance indicators. Default to FALSE}

\item{min.N.known.var}{minimum number of known SVs needed to build training model. Default to 50}

\item{bySVlength}{logical value, if true, performance will also be evaluated seperately for different categories of SV length. See details section for more}
}
\value{
a list contains DVboost.res which is essentially a \code{\link{gbm.object}} with several additional fields:
\itemize{
\item fitted.values: converted to probability based on \code{fit} field
\item ID: supplied IDs for SVs
\item is.known.variant: 0/1 indicating whether the SV is known(1) or novel(0)
\item DVboost.Q.score: Q scores for SVs
}
Besides, this will also save a set of figures and text files to \code{output.DIR.name}. Please see \code{\link{metricView}} for details
}
\description{
Building training models using subset of total SVs (Only 'DEL' type)
}
\details{
When bySVlength is set to true, training performance will be evaluated seperately for 5 different categories of SV length. To be specific,
the 5 SV length categories are 50-100bp, 100-500bp,500-1k, 1k-10k, >10k. For each SV length category, training performance will be evaluated and exported as
figures and text files.
}
\examples{
data(ExampleData, package='DVboost')
sample <- 'NA12878'
outdir <- getwd()
tmp.mtx.DEL <- ExampleData[ExampleData$SVType == 'DEL',]
truth.vec <- tmp.mtx.DEL$CNVMAP == 1 | tmp.mtx.DEL$CNVR ==1
is.semi.truth.vec <- as.numeric(truth.vec)
DVb.res <- runDVboostwrapper( var.atr.mtx = tmp.mtx.DEL, var.ID.vec = rownames(tmp.mtx.DEL),
                             is.known.var.vec = is.semi.truth.vec,
                             output.DIR.name = outdir,input.sample.ID=sample, bySVlength=FALSE)

}
\seealso{
\code{\link{metricView}}, \code{\link{fitDVboostmodel}}, \code{\link{DVboostQscore}}
}
